requirements.txt
absl-py==0.13.0
accelerate==1.0.1
addict==2.4.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.11
aiosignal==1.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.5.2
async-timeout==5.0.1
attrs==24.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
blis==0.7.11
blobfile==3.0.0
bvh==0.3
cachetools==4.2.2
catalogue==2.0.10
certifi==2021.5.30
charset-normalizer==2.0.4
chumpy==0.70
click==8.1.7
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
cloudpathlib==0.20.0
comm==0.2.2
confection==0.1.5
ConfigArgParse==1.7
cycler==0.10.0
cymem==2.0.10
dash==3.0.3
decorator==4.4.2
deepspeed==0.16.4
diffusers==0.31.0
einops==0.8.0
exceptiongroup==1.3.0
fastjsonschema==2.21.1
filelock==3.16.1
flask==3.0.3
frozenlist==1.5.0
fsspec==2024.2.0
ftfy==6.1.1
gdown==5.2.0
google-auth==2.36.0
google-auth-oauthlib==1.0.0
googletrans==4.0.2
grpcio==1.68.0
h11==0.16.0
h2==4.1.0
h5py==3.11.0
hjson==3.1.0
hpack==4.0.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.30.2
human-body-prior @ git+https://github.com/nghorbani/human_body_prior@4c246d8a83ce16d3cff9c79dcf04d81fa440a6bc
hyperframe==6.0.1
idna==3.2
imageio==2.9.0
imageio-ffmpeg==0.5.1
importlib-metadata==8.5.0
importlib-resources==6.4.5
ipdb==0.13.9
ipython==7.26.0
ipython-genutils==0.2.0
ipywidgets==8.1.6
itsdangerous==2.2.0
jedi==0.18.0
Jinja2==3.1.3
joblib==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter-core==5.7.2
jupyterlab-widgets==3.0.14
kiwisolver==1.3.1
langcodes==3.4.1
language-data==1.3.0
lightning-utilities==0.11.9
lxml==5.3.2
marisa-trie==1.2.1
Markdown==3.3.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.4.3
matplotlib-inline==0.1.2
mdurl==0.1.2
mkl-fft==1.3.0
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186064646/work
mkl-service==2.4.0
modelscope==1.23.1
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
murmurhash==1.0.11
narwhals==1.35.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.0
ninja==1.11.1.3
nltk==3.9.1
numpy==1.22.4
nvidia-cublas-cu11==11.11.3.6
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==9.1.0.70
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.3.0.86
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.5.86
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.570.86
nvidia-nccl-cu11==2.20.5
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.1.105
nvidia-nvtx-cu11==11.8.86
nvidia-nvtx-cu12==12.1.105
oauthlib==3.1.1
olefile==0.46
omegaconf==2.3.0
open3d==0.19.0
opencv-python==4.11.0.86
packaging==24.2
pandas==1.3.2
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
pillow==10.4.0
pkgutil-resolve-name==1.3.10
platformdirs==4.3.6
plotly==6.0.1
preshed==3.0.9
proglog==0.1.10
prompt-toolkit==3.0.20
propcache==0.2.0
protobuf==5.28.3
psutil==6.1.0
ptyprocess==0.7.0
py-cpuinfo==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodomex==3.22.0
pydantic==2.10.1
pydantic-core==2.27.1
pyDeprecate==0.3.2
pygments==2.18.0
pynvml==11.5.3
pyparsing==2.4.7
pyquaternion==0.9.9
PySocks==1.7.1
python-dateutil==2.8.2
pytorch-lightning==1.7.0
pytz==2021.1
PyYAML==5.4.1
referencing==0.35.1
regex==2024.11.6
requests==2.26.0
requests-oauthlib==1.3.0
retrying==1.3.4
rich==13.9.4
rpds-py==0.20.1
rsa==4.7.2
safetensors==0.4.5
scikit-learn==0.24.2
scipy==1.7.1
seaborn==0.13.2
sentence-transformers==3.2.1
sentencepiece==0.2.0
shellingham==1.5.4
six @ file:///tmp/build/80754af9/six_1623709665295/work
sklearn==0.0
smart-open==7.0.5
smplx==0.1.28
sniffio==1.3.1
soupsieve==2.6
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
sympy==1.13.1
tensorboard==2.14.0
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.0
tensorboardX==2.6.2.2
thinc==8.2.5
threadpoolctl==2.2.0
tokenizers==0.20.3
toml==0.10.2
torch==2.4.1+cu118
torchaudio==0.8.0a0+e4e171a
torchgeometry==0.1.2
torchmetrics==0.7.0
torchvision==0.9.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.46.3
triton==3.0.0
typer==0.13.1
typing-extensions==4.12.2
urllib3==1.26.6
wasabi==1.1.3
wcwidth==0.2.5
weasel==0.4.1
werkzeug==3.0.6
widgetsnbextension==4.0.14
wordcloud==1.9.4
wrapt==1.17.0
yarl==1.15.2
zipp==3.20.2
